perf(crypto): CRP-2962 Add affine Pippenger's implementation for BLS12-381
Already there was a function performing multiscalar multiplications taking as
input affine points, but previously this just converted the points into
projective format, as a result it was not able to take advantage of the
opportunity to use the cheaper mixed addition formulas.

With this change the benchmarked time to verify a NIDKG chunking proof is
reduced by about 5 milliseconds.

This also adds affine point variants of the mul2 functions. This allows using
mixed addition during the table setup phase.