perf(ckbtc): Reduce the number of instructions to build an unsigned transaction by a factor 5x to 600x (#7866)
Improve the efficiency of the ckBTC minter when handling withdrawals,
where the main cost driver in terms of number of instructions is lead by
finding a suitable subset of UTXOs that can be used in a Bitcoin
transaction. Using the benchmarks introduced in #7782, the number of
instructions required to prepare an unsigned transaction is reduced by a
factor 5 to more than 600 on _mainnet_, depending on the withdrawal
amount.

The main optimization consists of storing the UTXOs being sorted by
value, instead of the canonical order defined on
[`ic-btc-interface::Utxo`](https://docs.rs/ic-btc-interface/0.2.3/src/ic_btc_interface/lib.rs.html#298)
consisting of the lexicographic order on `OutPoint`.

| Bench | Before | After | Reduction factor |

|----------------------------------------|--------------|----------|------------------|
| build_unsigned_transaction_1_50k_sats | 14,675,909 | 2,691,537 | 5 |
| build_unsigned_transaction_2_100k_sats | 14,675,757 | 2,691,629 | 5 |
| build_unsigned_transaction_3_1m_sats | 13,738,959 | 2,692,824 | 5 |
| build_unsigned_transaction_4_10m_sats | 14,282,438 | 2,693,220 | 5 |
| build_unsigned_transaction_5_1_btc | 416,462,644 | 3,121,724 | 133 |
| build_unsigned_transaction_6_10_btc | 4,916,848,618| 7,842,419 | 626 |