feat(registry): CRP-2618 allow omitting pre_signatures_to_create_in_advance for keys that don't require it (#7859)
Adapts the registry canister so that during subnet creation, subnet
update, and subnet recovery, the `KeyConfig`'s field
`pre_signatures_to_create_in_advance` can be omitted for keys that do
not have pre-signatures. Currently the only keys that do not require
pre-signatures are VetKD keys (as opposed to Ecdsa/Schnorr keys). When
the field is omitted, it is automatically set to zero. So far, because
this field was always required, it was set to 0 manually for vetKD keys
as part of proposal creation.

Note that there is a registry invariant that requires this field to be
non-zero for keys that do have pre-signatures. For keys that do not have
pre-signatures, any value is permitted (i.e., there is no invariant).
The replica simply ignores this field for vetKD.

This PR is a first step towards making it mandatory to omit the
`pre_signatures_to_create_in_advance` field for vetKD keys. This step is
necessary because our CI runs the system tests both agains the head NNS
and the mainnet NNS, and directly omitting this field in the replica
would make the system tests that run agains mainnet NNS fail.