chore: remove the `local` tag from ic_icrc_rosetta_image (#7921)
This removes the `local` tag from the `oci_image`
`//rs/rosetta-api/icrc1:ic_icrc_rosetta_image` such that the image gets
cached and build in a sandbox.

In the PR that introduced it there was the
[comment](https://github.com/dfinity/ic/pull/4569/files#r2021000986):
> We are changing to `local` because it provides more context on how it
was built.

But I don't see what context gets embedded in the image that's different
between building in a sandbox and building outside of it. If there's
important context that needs to be in the image it should be provided as
an explicit input to the rule.