test: CON-1577 Pre-signature stash management system test (#6616)
Any ECDSA/Schnorr pre-signatures that were stored in replicated state
should be purged whenever the key transcript is rotated. Such key
rotations happen periodically, during membership changes, and subnet
recoveries.

This PR extends the existing key rotation and app subnet recovery tests,
to verify that the pre-signature stash is indeed purged correctly. This
is done by disabling pre-signature generation before the key rotation,
and then asserting that the stash is empty, after the rotation was
completed.

Additionally, this PR adds a new system test to test the management of
the pre-signature stash, in general: It should be possible to grow and
shrink the maximum pre-signature stash size on demand by adopting a
governance proposal.

Note that for now, the new test functionality is still disabled behind
the `STORE_PRE_SIGNATURES_IN_STATE` flag, as is the actual
implementation of the feature. Enabling the feature will thereby also
enable the new system test cases.