fix(crypto): use correct vault type when building temp crypto (#7869)
The temp crypto component test utility has a bug in its builder that
makes it always build the crypto component with the _local_ vault,
regardless of whether the builder was configured with
`with_remote_vault` to use the _remote_ vault. The intermediate layer,
the crypto service provider (CSP) layer, is built with the correct vault
(namely, the one that was configured in the builder, with the local one
being the default one), which means this only affects tests and
benchmarks that make use of crypto component functionality that calls
the vault directly rather than via the CSP layer. Such functionality is
iDKG/tECDSA.

This PR fixes this, by using the correct vault type according to the
builders configuration: the local vault by default, and the remote vault
in case `with_remote_vault` was used.