feat: adding a QR code encoding for a node id during registration (#7930)
This PR starts a sequence of orchestrator required changes for the new
registration process that has a goal to make managing the node operator
key more securely and to allow for better user experience during node
installations.

At the moment, the orchestrator won't print a QR code on actual
setupOS's built for mainnet since it is still not required. In the
future PR's modifications to configuration will be made and the
displaying of the qr code will be an opt in feature when the other
components are ready.