perf(crypto): CRP-2965 Avoid adding into the identity element
In various elliptic curve point multiplication algorithms, an
accumulator element is created (initialized as the identity element)
and then several points are added to it.

However the first point addition is always completely avoidable; it's
equvalent to just an assignment.

Restructure the multiplication algorithms such that the accumulator is
instead initialized with the value created in the first loop of the
multiplication algorithm.