fix: drop unnecessary DefaultInfo in release_nostrip (#7965)
The `release_nostrip_binary` rule would copy the `DefaultInfo` of the
executable (binary) being transitioned. This is not entirely correct and
creates issues in Bazel 8. Moreover, nothing but the resulting release
binary should be necessary, so we can resolve the issue by specifying
`executable` only.

Additionally, the transition function is renamed for consistency.