feat: add support for CRITERION_HOME
This introduces a new CLI argument `--criterion-home` that we intercept
and use to set `CRITERION_HOME` for the bench runner. See comment in
code for details.

This allows us to specify the criterion root (home) where we start
looking for bench dirs, in order to avoid false positives from
`bazel-out/`.