perf(ckbtc): avoid cloning the UTXOs set in `estimate_withdrawal_fee` (#7974)
Improve the efficiency of the endpoint `estimate_withdrawal_fee` by
removing the cloning of the UTXO set, which can be quite large (around
60k entries for ckBTC).

The cloning is not necessary because the endpoint
`estimate_withdrawal_fee` is a query endpoint, so that event though the
underlying logic (the methods `estimate_retrieve_btc_fee` and
`estimate_retrieve_doge_fee`) needs a mutable reference to the UTXO set
(to remove selected UTXOs), any state change will be reverted. This PR
also introduces a benchmark for the ckBTC minter endpoint to measure the
exact gains.

| Bench | Before | After | Reduction factor |

|----------------------------------------|--------------|----------|------------------|
| build_estimate_retrieve_btc_fee_1_50k_sats | 40,813,570 | 2,689,868 |
15 |