refactor(migration-canister): cache last event for every migration (#7948)
This PR refactors the stable memory representation of events in the
migration canister to
- the history of all events in a map indexed by their sequence numbers;
- the last event for every pair of source and target canisters
represented by its index into the history.

Alternatives considered:
- a single map `BTreeMap<CanisterMigrationArgs, Vec<Event>, Memory>` -
this representation is inefficient if there are (maliciously) many
events for a given pair of source and target canisters;
- a single map `BTreeMap<Event, (), Memory>` with events sorted by their
pairs of source and target canisters - relying on range queries is
fragile and introduces substantial complexity for testing.

---------

Co-authored-by: michael-weigelt <122277901+michael-weigelt@users.noreply.github.com>