feat: lazy initialization for log_memory_store to avoid unnecessary log-memory charges (#7982)
This PR introduces a lazy initialization for `log_memory_store` to avoid
charging for dirty page_map.

The canister can be created without wasm module or be uninstalled (when
all its memories are cleaned). In those states the canister should
require no charge, therefore should use zero memory (especially in
page_map's). Previously we were storing `log_memory_limit` implicitly in
ring-buffer header in the first page of page_map, which dirtied the page
and required memory usage charging even if canister had no wasm module
or was uninstalled.

In this PR we store `log_memory_limit` outside of page_map keeping it
clean and only initizlizing it when appending logs. In the follow up PR
this value will be preserved in canister state bits protobuf.