test: [DSM-50] Execution tests for online_split()
Duplicate the existing Execution tests for after_spit() (testing that in-progress stop_canister(), install_code() and raw_rand() requests for migrated canisters are rejected) for the newly introduced online_split(). The existing tests for after_split() will be dropped soon together with split() and after_split(), but for the time being, we still require the MVP split functionality, just in case.