perf(crypto): CRP-1559 Use affine rather than projective points for BLS12-381 signture protocols (#7969)
The `multi_sig` and `threshold_sig` BLS12-381 signature libraries used
projective points for public keys and signatures. Various operations
such as serialization or verifying a BLS signature require that the
input points are in affine format, and converting from projective to
affine is relatively expensive. Other operations like point addition are
cheaper if one of the elements is in affine form. Also when
deserializing a point the natural encoding is affine, not projective. So
for example when verifying a multi signature, the signature would be
presented to the crypto component as bytes, deserialized into projective
form, then converted back to affine in order to verify the signature
equation.

Serialization and deserialization of the relevant artifacts is
unaffected. Projective points cannot be serialized directly, instead
they must anyway be converted into affine form, which right now happens
internally in `G(1|2)Projective::serialize`. Likewise, deserialization
is always into affine coordinates, and then the point is converted to
projective.

This removes serialization and deserialization functions from the
projective types, as potential performance footguns. For serialization,
that's because the serialization hides the (expensive) conversion to
affine form. For deserialization, it's because keeping the point in
affine form as long as possible is typically best for performance.