chore(node): improve manual recovery fallback logging (#8009)
<img width="1300" height="975" alt="image"
src="https://github.com/user-attachments/assets/15632699-80ad-4246-b7db-31fb63bd6d52"
/>

Add a success log in guestos-recovery-upgrader.sh. This way, when the
manual recovery fallback mechanism is used (running
guestos-recovery-launcher.sh directly), a success message is still
logged to screen.

Note that when the recovery TUI is used, a success message is printed
_after_ the recovery console is closed. This way, the success message
persists to screen after the console is wiped.

So when the recovery TUI is now used (the primary recovery method),
**two** successive success messages are logged, but the first one comes
before the console is wiped, so only one still persists to screen.