refactor: Simplify snapshot checkpointing (#7998)
This commit makes checkpointing for canister snapshots more robust, by
removing some cases from `UnflushedCheckpointOps`. This makes it harder
to make inconsistent changes to `ReplicatedState` that could lead to
problems such as dropped data. A problem like that was recently fixed in
commit 86c4ed2de50c078a309494adffe9eb3bc3099f01. This refactor aims at
making it robust enough so that such fixes are not needed in the future.

This commit contains the following changes:
1. Dropping `UnflushedCheckpointOps::UploadSnapshotMetadata` and
`UnflushedCheckpointOps::UploadSnapshotData`: Previously, only snapshots
with one of these unflushed checkpoint ops would be checked for changes
during checkpointing and persisted. Instead, the checkpointing logic now
iterates over all snapshots. This is consistent with how it handles
changes to canister states, which are also iterated during
checkpointing.

This is less efficient, but more robust. It is no longer possible to
change a canister snapshot in a way that would then not be picked up by
checkpointing. The performance penalty is manageable, considering we
already do something similar for canisters.
2. Dropping `UnflushedCheckpointOps::DeleteSnapshot`: Again, we align
the implementation with how canisters are handled. Instead of relying on
the unflushed checkpoint op, we instead compare the list of snapshots
with the files on disk, detecting snapshots that have been deleted in
the meantime. This makes it harder to miss a deleted snapshot due to
inconsistent checkpoint ops, while again being less efficient.

Note that the checkpoint ops `TakeSnapshot`, `LoadSnapshot` and
`RenameSnapshot` are still present and necessary. They contain extra
information such as what canister a snapshot is restored to. This
information cannot be deduced otherwise by simply looking at the state
during checkpointing. Furthermore, the exact order of these operations
is relevant for checkpointing as well.