refactor(crypto): CRP-2666 remove CSP layer for basic-sig signing (#7966)
Adapts the crypto component so that the creation of basic (=Ed25519)
signatures is done by directly calling the vault, rather than doing the
obsolete indirection via the crypto service provider (CSP) layer.

With this PR, basic signature signature creation also no longer
calculates a (private) key ID from the signing node's _node signing_
public key stored in the registry, and thus also no longer reads this
public key from the registry.

Also removes the old set of basic-signature unit tests and creates a
fresh set of respective integration tests.

This also makes the `BasicSigner`s `signer: NodeId` and
`registry_version: RegistryVersion` parameters obsolete/unused, which
will be removed in a follow-up PR.

The respective code for basic-signing in the CSP is kept for now,
because it is still used in various tests for basic-sig verification in
the CSP, and can/will be removed once we also go directly to the vault
for basic-sig verification.

---------

Co-authored-by: IDX GitHub Automation <infra+github-automation@dfinity.org>