chore(DSM): Make backtrace tests robust to compiler changes
Our tests on canister backtraces currently assert that error messages
contain exact backtraces for test canisters built with Rust. But those
backtraces can change slightly with different versions of the Rust
compiler or the CDK.

This makes those tests more robust to changes in the compiler or CDK by
performing regex matches against the essential parts of the expected
backtrace. This allows us to ignore boilerplate inserted by the compiler
or CDK.