feat(ICRC_Ledger): DEFI-2541: Manual Block deserializer implementation (#7847)
As preparation for supporting the parsing of blocks that have no `tx.op`
field in the ICRC-3 block, this PR proposes to manually implement the
`Deserialize` trait for `Block`. Internally, the deserialization
implementation then uses both the `btype` of the `Block`, as well as the
parsed `FlattenedTransaction`, to parse the `Transaction` that is
subsequently used e.g., by the index to apply the transaction in the
block.