chore: More detailed state sync download metrics
Previously, the metric `state_sync_manager_chunk_download_duration`
would combine the download with the call to `add_chunk`, which does
varying amounts of work depending on which phase of the state sync we
are in. This made it for example difficult to distinguish a node with a
slow network from a node with a slow disk.

This commit changes `state_sync_manager_chunk_download_duration` to only
consider the download, and introduces a new metric
`state_sync_add_chunk_duration_seconds` for the more disk-heavy call to
`add_chunk`. Furthermore, the new metric is broken down by the status of
the state sync.