fix(ic-boundary): start polling the registry client (#8045)
https://github.com/dfinity/ic/pull/5909 made the registry replicator
internally take care of keeping its integrated registry client in sync.
That PR thus removed the polling logic of the registry client in
`ic-boundary` since it assumed that the registry replicator would do it
itself. Though, the registry replicator is only ran when
`cli.registry.registry_disable_replicator` is false, otherwise the
orchestrator takes care of replicating the registry to disk. In that
second case, it was overlooked that the registry client inside the
`ic-boundary` process would thus not be updated.

This PR reverts most changes of that PR on `ic-boundary`, i.e., creates
a registry client and starts polling it independently of whether we will
start a replicator. This is also in sync with what the replica
[does](https://github.com/dfinity/ic/blob/bc684d1e033376dffeacda7801cab2726528ba0d/rs/replica/src/setup.rs#L194-L196):
in both cases, those registry clients are in a different process than
the replicator (which lives in the orchestrator), so we cannot ask the
replicator to manage those clients.

In theory, you could also imagine that in the case where we spawn a
replicator in `ic-boundary`, we could reuse its internal registry client
instead of having a separate one like I propose, but I think this would
add some added complexity to the code for a scenario for which I am not
sure is common (do we ever start `ic-boundary` without an orchestrator?)

Note that the original functionality that would panic if the local store
is uninitialized and we do not have an NNS public key in the config is
kept since it is now part of the [replicator's
constructor](https://github.com/dfinity/ic/blob/bc684d1e033376dffeacda7801cab2726528ba0d/rs/orchestrator/registry_replicator/src/lib.rs#L109-L115).