feat(ICRC-Rosetta): DEFI-2564: add schema versioning (#8041)
We add a `rosetta_metadata` table and store the `schema_version` in the
table. If rosetta is started with schema version that is higher than its
current schema version, it reports an error on startup. This can happen
if the database file is created by a newer version of rosetta and
rosetta is downgraded to an older version. If rosetta proceeded without
an error, it could miss or misinterpret some of the stored data and
report incorrect balances.