fix: [CON-1553] orchestrator uses correct IC root key in testnet (#5841)
The orchestrator uses `ic-agent` since #5686. By default, the agent uses
the mainnet's root key to verify responses. This means that the
orchestrator could not verify responses in testnet. All tests still pass
because the requests actually go through, just that the orchestrator
cannot verify the responses.

This PR fixes this by setting the root key in the agent to:
- The one in the config for node registration.
- The one from the registry, falling back to the config, for IDKG key
rotation.

Additionally, the PR also refactors the registration loop a bit to avoid
some leftovers panics and make it more readable by moving everything to
its own function.

Finally, the PR removes the code duplication in
`register_node`/`is_node_registered`, which also avoids logging `Node
keys are not setup: ...` twice.