feat(ICRC_Ledger): DEFI-2541: Make tx.op string optional, but still require Operation in Transaction (#7848)
As a follow-up to [PR 7847](https://github.com/dfinity/ic/pull/7847),
add support for other `tx.op` values, in particular for the ICRC-107
`107feecol` block. This PR proposes to explicitly add the string value
for the `tx.op` field to the `FeeCollector` variant of the `Operation`
enumeration, and decouple the `tx.op` value from the `Operation`
enumeration variant the way it is implemented for the legacy ICRC-1 and
ICRC-2 operations, where the `tx.op` value is always fixed.

---------

Co-authored-by: IDX GitHub Automation <infra+github-automation@dfinity.org>