chore(PocketIC): simplify and decouple types between server and client (#7936)
This PR simplifies and decouples the configuration types for PocketIC
"ticks" between the server and client (user-facing library). While the
simplification and decoupling is technically a breaking change, we
expect its impact to be limited since the current configuration options
(configuring subnet blockmakers) are only expected to be used
internally.