chore(migration-canister): refine migrated/replaced with "canister" and fix typos (#8059)
This PR refines occurrences of "migrated" or "replaced" in the migration
canister with "canister" to avoid ambiguity and fixes a typo in the
public API:
-
[d57f4f5](https://github.com/dfinity/ic/commit/d57f4f5abaae0a5cb3623d6683aa7fd7e3e91ef3):
`replace_canister_id` -> `replaced_canister_id` in the input type of the
public `migrate_canister` endpoint;
-
[c50f5fe](https://github.com/dfinity/ic/commit/c50f5fec90b362abad494aff4aacf8961d1912c2):
append `Canister` to `Migrated` and `Replaced` in `ValidationError`
variants;
-
[4b50428](https://github.com/dfinity/ic/commit/4b5042845656bed9073dc9c09d5c6021df97e14e):
append `Canister` to `Migrated` and `Replaced` in `RequestState`
variants;
-
[6c0cace](https://github.com/dfinity/ic/commit/6c0cacec76a301c79b0e0125efc1325a9abd6c22):
add `_canister` to `migrated` and `replaced` in variable names, internal
`CanisterMigrationArgs` type, and internal `Request` type;
-
[72a1986](https://github.com/dfinity/ic/commit/72a1986b6e2f1da37f2e63ed072907edc5f39d10):
ad-hoc fixes (debug message in an assert and timer predicate
identifier).

The changes from this PR are breaking changes and require a reinstall of
the canister in production!