refactor(ckbtc): Generic events (#7991)
Refactor the ckBTC event log to be generic over the type of events
stored, so that the ckDOGE minter, that reuses code from the ckBTC
minter can have dedicated events. Quite a few ckBTC events are only
there for backwards-compatibility reasons or are not relevant for the
ckDOGE minter (e.g. events related to checking whether UTXOs or
addresses are tainted).

This PR is pure refactoring and does not change the behaviour of the
ckBTC or ckDOGE minter. It only allows to have 2 implementations, which
are currently identical, on how to record and replay events. The next PR
will change the ckDOGE implementation to have dedicated events.

1. #7991
2. #8035

---------

Co-authored-by: Thomas Locher <thomas.locher@dfinity.org>
Co-authored-by: IDX GitHub Automation <infra+github-automation@dfinity.org>