fix(ICP-Rosetta): fix flaky test_disburse_maturity test (#8083)
The flakiness is hard to reproduce, but the failure is
`SettingTimeIntoPast((1766416868853434997, 1766416868173763543))` and
happens during advancing time. Since this is less that 1s, it seems that
advancing time by 1s might be flaky when system is under high load - by
the time the time is advanced, more than 1s passed. Since advancing time
by more than 1s also works for the test, I changed it to 60s.