refactor(lifeline): Spin out tests in rs/nns/handlers. (#7993)
Basically, what I want all of our code to do is this:

```
#[cfg(test)]
#[path = "foo_tests.rs"]
mod tests;
```

instead of inlining.

(In the special case of lib.rs, no need for `#[path ...]`.)

I think inlining is actually currently our predominant way, but I don't
think that was intentional. It was just the "natural" way/path of least
resistance. It maybe didn't make a difference originally when we only
had cargo, but with bazel, this has significant potential to speed up
development.

Root was already following the above pattern. Therefore, only `lifeline`
is changed here.

Also, got rid of the `lifeline_bin` rule, because it doesn't seem to be
used, but this is orthogonal.

# Why the Pattern?

This speeds up build times when you change a test. When you only change
a test, it should not be necessary to re-build anything other than the
test, but when tests are inlined, you ALSO invalidate the rust_library
result, because an input .rs file is change. In other words, when tests
are inlined, you are trashing the build cache.

The other piece that you need to make this optimization work is in
`BUILD.bazel`:

```
LIB_SRCS = glob(
    ["src/**/*.rs"],
    exclude = ["**/*tests.rs"],
)

rust_library(
    ...
    srcs = LIB_SRCS,
    ...
)
```

Another advantage is that when you want search results to exclude tests,
shunting test code to *tests.rs files makes it easy to have cleaner
search results.