feat: [DSM-50] Subnet splitting in the DSM (#7836)
Implements "online subnet splitting" within the DSM, triggered by an
explicit instruction from Consensus.

"Online subnet splitting" is triggered in a special checkpoint round.
Consensus provides us with an updated registry and each replica is
instructed which of the two subnets it is now part of ("subnet A'" or
"subnet B").

All paused executions are aborted but no other message execution or
routing happens; the relevant parts of the state are split (canisters,
snapshots), the rest of it is retained on the original subnet ("subnet
A'") and reset on the newly created subnet ("subnet B"); then distinct
checkpoints are produced by the two sets of replicas.

Finally, Consensus produces two CUPs, one from each set of replicas.
Then the replicas are restarted, as two separate subnets.