fix: Fix HostOS machine-id (#8088)
The set-machine-id script was previously added to a subdirectory under
`/etc/initramfs-tools/scripts/init-bottom` so `update-initramfs` didn't
pick it up, so it was never included in the initramdisk. Even if it had
been included, the script tried to use the non-existing `/dev/vda3`
(likely GuestOS copy-paste) to find the machine-id.