feat: [DSM-50] Introduce a new subnet split marker (#7999)
The existing split marker field (`split_from`) both triggers the
execution of phase 2 of splitting (`after_split()`) and is persisted in
a separate checkpoint file (so as to not change the hash of the metadata
state checkpoint file). Introduce a new `subnet_split_from` field,
persisted with the rest of metadata state, that only causes the
`mr_subnet_split_height` metric to be exported and is then reset.