feat(DSM-58): Refactoring for Deterministic Memory Tracker (#8095)
[DSM-58](https://dfinity.atlassian.net/browse/DSM-58)

Refactoring of the `MemoryTracker` trait and `memory_tracker` crate to
simplify the changes which introduce the deterministic memory tracker.
The changes are:

1. Modify the `MemoryTracker` trait to include methods returning the
internal `MemoryArea` and metrics instead of having individual functions
which get passed to those internal values.
2. Move some code from the `prefetching` module to the crate top level
because it will be also be used for the deterministic tracker.
3. Modify existing tests to use a `dyn MemoryTracker` instead of the
`PrefetchingMemoryTracker` so that they can be run on the deterministic
tracker as well.

[DSM-58]:
https://dfinity.atlassian.net/browse/DSM-58?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

---------

Co-authored-by: IDX GitHub Automation <infra+github-automation@dfinity.org>