feat(IDX): reduce max bazel-remote connections (#8096)
We've seen a lot of these errors recently:
```
DEADLINE_EXCEEDED: context deadline exceeded
Error while uploading artifact
```
The hypothesis is that the uploads timeout for large files due to too
many uploads competing for resources. We want to try reducing the number
of total connections to see if it can enable higher throughput.

Note that `--remote_max_concurrency_per_connection` is only available
for bazel 8, so we can't set it yet.

My suggestion is to start with half of the amount of connections we've
had before. So far I have seen no `context deadline exceeded` messages
in this PR (after testing with various changes). Though I think we'll
need to merge to master to measure the impact.

---------

Co-authored-by: IDX GitHub Automation <infra+github-automation@dfinity.org>