chore: simplify `ict` by running `bazel run` instead of `bazel test` (#8087)
For simplification we're working towards deprecating the `ict` tool and
just relying on `bazel` alone.

For this we simplify the `ict testnet create <testnet>` command to
invoke
`bazel run <testnet>` instead of
`bazel test <testnet>`.

The former bazel invocation will now keep the testnet alive instead of
doing this from `ict`. This does mean that the required `ict`
`--lifetime-mins=N` option is no longer supported. This required option
would limit the lifetime of the testnet to the specified number of
minutes. Now `ict testnet create <testnet>` will keep the testnet alive
as long as the `ict` command is running.