chore: add log_memory_limit to canister system_state and checkpoint proto (#8120)
This PR returns back the `log_memory_limit` to proto and `system_state`.

It was previously removed in https://github.com/dfinity/ic/pull/7922
with the idea that this value is going to be stored in
`log_memory_store` field, which is a PageMap-ped ring-buffer and needed
no duplication. But it turned out there's a special case where the
canister is created but no code installed and therefore it should
preserve the setting of `log_memory_limit` without actually allocating
memory for ring-buffer. Therefore we should store this field separately
both in `system_state` and preserve in proto to survive the checkpoint.

The difference to https://github.com/dfinity/ic/pull/7922 is it's not
using hardcoded default values, because they will be overwritten later
by applying default canister settings.