feat: match exact targets just by target name if they are unique (#8143)
This PR is a quality of life improvement that improves how ict matches
target names if less than a full path is given.

Basically since each target has the following targets (some of them have
just the colocated one which was already handled):
```bash
//rs/tests/nested:registration
//rs/tests/nested:registration_colocate
//rs/tests/nested:registration_head_nns
//rs/tests/nested:registration_head_nns_colocate
```
The tool will match it by exact match of the full name. So If the user
runs
```bash
ict test registration
```
It would match `//rs/tests/nested:registration` instead of saying there
are 4 targets and it doesn't know which one to use.