chore: extract script from run_system_test rule (#8157)
This moves some of the implementation of `run_system_test` out of the
custom rule in preparation for replacing it with a macro.

This also adds documentation to some of the run script's inputs,
simplifies some of the logic and extracts the run script into a
dedicated bash file for automatic formatting and linting and ease of
editing. The systest test "remote dep" uploader is also moved to the
system test directory for consistency.

The run script's main command is now also `exec`ed to avoid nesting
processes.

---------

Co-authored-by: Bas van Dijk <bas@dfinity.org>