test(mainnet-nns): reintroduction of the mainnet NNS testnet (#8170)
This PR introduces a new testnet `mainnet_nns`: A single-node NNS subnet
holding mainnet state, along with 1 API BN and 1 HTTP Gateway, for
general use-case. In a separate PR, the current `nns_recovery` testnet
will be adapted to also be able to spawn it with mainnet state. Finally,
that second PR will also introduce an NNS recovery system test that will
use mainnet state.

Most of the changes are the revert of
https://github.com/dfinity/ic/commit/e38941af541aaa4b15b5d5676872077e6010d1ca,
but some changes were required in `ic-replay`:
- Bumping the ingress messages' expiry, otherwise some messages would
expire if delivered in multiple batches. This was already done for
certain `SubCommand`s, so we generalize it to all of them.
- The list of trusted neurons was outdated.
- Now neurons need to be made public to be followed.
- Trusted neurons will follow the new neuron on all topics and not just
`Unspecified` to be more robust.

The code lives in its own crate in `//rs/tests/testnets/mainnet_nns/` as
a library such that one can turn any testnet or system test into a
variant that holds mainnet state, with as few modifications as possible
to their original code. For example, the setup of the IC can be done by
running the provided `setup` and making proposals to change the topology
to achieve the intended initial topology. Such proposals can be made as
easily as calling `ProposalWithMainnetState::[name_of_the_proposal]`.

Finally, it is intended to use the mainnet's version of `ic-replay` and
`ic-recovery` like
[previously](https://github.com/dfinity/ic/blob/6cd54499b0ec4f4190043f66769e54e78ffa0ccf/rs/tests/nns/ic_mainnet_nns_recovery/src/lib.rs#L206-L212),
but since I needed to modify those to create this testnet, we need to
wait for this change to reach mainnet NNS before downloading those. This
is tracked in ticket CON-1624.

---------

Co-authored-by: IDX GitHub Automation <infra+github-automation@dfinity.org>